# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Water Plant CRM API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProductionRecord {
  id: number;
  date: string;
  litersFiltered: number;
  chemicalMixed: number;
  wasteWater: number;
  cansFilled: number;
  cansDelivered: number;
  cansLeft: number;
  waterLeft: number;
  machineRunningHours: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ProductionInput {
  date: string;
  litersFiltered: number;
  chemicalMixed: number;
  wasteWater: number;
  cansFilled: number;
  cansDelivered: number;
  machineRunningHours: number;
  notes?: string;
}

export interface ProductionUpdate {
  litersFiltered?: number;
  chemicalMixed?: number;
  wasteWater?: number;
  cansFilled?: number;
  cansDelivered?: number;
  machineRunningHours?: number;
  notes?: string;
}

export interface Expense {
  id: number;
  date: string;
  category: string;
  amount: number;
  description: string;
  createdAt: string;
}

export interface ExpenseInput {
  date: string;
  category: string;
  amount: number;
  description: string;
}

export interface ExpenseUpdate {
  date?: string;
  category?: string;
  amount?: number;
  description?: string;
}

export interface CategoryTotal {
  category: string;
  total: number;
}

export interface ExpenseSummary {
  month: string;
  totalExpenses: number;
  byCategory: CategoryTotal[];
}

export interface IncomeRecord {
  id: number;
  date: string;
  cashAmount: number;
  upiAmount: number;
  creditAmount: number;
  totalAmount: number;
  customerName: string;
  liters: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface IncomeInput {
  date: string;
  cashAmount: number;
  upiAmount: number;
  creditAmount: number;
  customerName: string;
  liters: number;
  notes?: string;
}

export interface IncomeUpdate {
  date?: string;
  cashAmount?: number;
  upiAmount?: number;
  creditAmount?: number;
  customerName?: string;
  liters?: number;
  notes?: string;
}

export interface MonthlyStats {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  cashIncome: number;
  upiIncome: number;
  creditIncome: number;
  totalLiters: number;
  totalCansDelivered: number;
}

export interface MonthlyDashboard {
  months: MonthlyStats[];
}

export interface TodayDashboard {
  date: string;
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  cashIncome: number;
  upiIncome: number;
  creditIncome: number;
  litersFiltered: number;
  cansFilled: number;
  cansDelivered: number;
  cansLeft: number;
  machineRunningHours: number;
}

export type ListProductionParams = {
/**
 * Filter by date (YYYY-MM-DD)
 */
date?: string;
/**
 * Filter by month (YYYY-MM)
 */
month?: string;
};

export type ListExpensesParams = {
/**
 * Filter by month (YYYY-MM)
 */
month?: string;
category?: string;
};

export type GetExpenseSummaryParams = {
/**
 * Month filter (YYYY-MM)
 */
month?: string;
};

export type ListIncomeParams = {
/**
 * Filter by month (YYYY-MM)
 */
month?: string;
};

export type GetMonthlyDashboardParams = {
/**
 * Number of months to include (default 6)
 */
months?: number;
};