# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Water Plant CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  date: string;
  category: string;
  amount: number;
  description: string;
  createdAt: string;
}