# lib/api-zod/src/generated/types/expenseSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Water Plant CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryTotal } from './categoryTotal';

export interface ExpenseSummary {
  month: string;
  totalExpenses: number;
  byCategory: CategoryTotal[];
}