# lib/api-zod/src/generated/types/incomeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Water Plant CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface IncomeInput {
  date: string;
  cashAmount: number;
  upiAmount: number;
  creditAmount: number;
  customerName: string;
  liters: number;
  notes?: string;
}