# lib/api-zod/src/generated/types/monthlyDashboard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Water Plant CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyStats } from './monthlyStats';

export interface MonthlyDashboard {
  months: MonthlyStats[];
}