# lib/api-zod/src/generated/types/monthlyStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Water Plant CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyStats {
  month: string;
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  cashIncome: number;
  upiIncome: number;
  creditIncome: number;
  totalLiters: number;
  totalCansDelivered: number;
}